requirements.txt
requests
django
djangorestframework
